requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.2
Pillow==9.5.0
tqdm==4.65.0
matplotlib==3.7.1